Split out ToolsPanelDropdownMenu
