+++ packages/components/src/tools-panel/tools-panel-dropdown-menu/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
