+++ packages/components/src/tools-panel/tools-panel-header/component.tsx
@@ -6,223 +6,45 @@ import type { ForwardedRef } from 'react';
 /**
  * WordPress dependencies
  */
-import { speak } from '@wordpress/a11y';
-import { check, reset, moreVertical, plus } from '@wordpress/icons';
 import { __, _x, sprintf } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import DropdownMenu from '../../dropdown-menu';
-import MenuGroup from '../../menu-group';
-import MenuItem from '../../menu-item';
 import { HStack } from '../../h-stack';
 import { Heading } from '../../heading';
 import { useToolsPanelHeader } from './hook';
 import { contextConnect, WordPressComponentProps } from '../../ui/context';
-import type {
-	ToolsPanelControlsGroupProps,
-	ToolsPanelHeaderProps,
-} from '../types';
-
-const DefaultControlsGroup = ( {
-	items,
-	toggleItem,
-}: ToolsPanelControlsGroupProps ) => {
-	if ( ! items.length ) {
-		return null;
-	}
-
-	return (
-		<MenuGroup>
-			{ items.map( ( [ label, hasValue ] ) => {
-				if ( hasValue ) {
-					return (
-						<MenuItem
-							key={ label }
-							role="menuitem"
-							icon={ reset }
-							label={ sprintf(
-								// translators: %s: The name of the control being reset e.g. "Padding".
-								__( 'Reset %s' ),
-								label
-							) }
-							onClick={ () => {
-								toggleItem( label );
-								speak(
-									sprintf(
-										// translators: %s: The name of the control being reset e.g. "Padding".
-										__( '%s reset to default' ),
-										label
-									),
-									'assertive'
-								);
-							} }
-						>
-							{ label }
-						</MenuItem>
-					);
-				}
-
-				return (
-					<MenuItem
-						key={ label }
-						role="menuitemcheckbox"
-						icon={ check }
-						isSelected
-						aria-disabled
-					>
-						{ label }
-					</MenuItem>
-				);
-			} ) }
-		</MenuGroup>
-	);
-};
-
-const OptionalControlsGroup = ( {
-	items,
-	toggleItem,
-}: ToolsPanelControlsGroupProps ) => {
-	if ( ! items.length ) {
-		return null;
-	}
-
-	return (
-		<MenuGroup>
-			{ items.map( ( [ label, isSelected ] ) => {
-				const itemLabel = isSelected
-					? sprintf(
-							// translators: %s: The name of the control being hidden and reset e.g. "Padding".
-							__( 'Hide and reset %s' ),
-							label
-					  )
-					: sprintf(
-							// translators: %s: The name of the control to display e.g. "Padding".
-							__( 'Show %s' ),
-							label
-					  );
-
-				return (
-					<MenuItem
-						key={ label }
-						icon={ isSelected && check }
-						isSelected={ isSelected }
-						label={ itemLabel }
-						onClick={ () => {
-							if ( isSelected ) {
-								speak(
-									sprintf(
-										// translators: %s: The name of the control being reset e.g. "Padding".
-										__( '%s hidden and reset to default' ),
-										label
-									),
-									'assertive'
-								);
-							} else {
-								speak(
-									sprintf(
-										// translators: %s: The name of the control being reset e.g. "Padding".
-										__( '%s is now visible' ),
-										label
-									),
-									'assertive'
-								);
-							}
-							toggleItem( label );
-						} }
-						role="menuitemcheckbox"
-					>
-						{ label }
-					</MenuItem>
-				);
-			} ) }
-		</MenuGroup>
-	);
-};
+import type { ToolsPanelHeaderProps } from '../types';
+import ToolsPanelDropdownMenu from '../tools-panel-dropdown-menu';
 
 const ToolsPanelHeader = (
 	props: WordPressComponentProps< ToolsPanelHeaderProps, 'h2' >,
 	forwardedRef: ForwardedRef< any >
 ) => {
-	const {
-		areAllOptionalControlsHidden,
-		dropdownMenuClassName,
-		hasMenuItems,
-		headingClassName,
-		label: labelText,
-		menuItems,
-		resetAll,
-		toggleItem,
-		...headerProps
-	} = useToolsPanelHeader( props );
+	const { headingClassName, label, resetAll, toggleItem, ...headerProps } =
+		useToolsPanelHeader( props );
 
-	if ( ! labelText ) {
+	if ( ! label ) {
 		return null;
 	}
 
-	const defaultItems = Object.entries( menuItems?.default || {} );
-	const optionalItems = Object.entries( menuItems?.optional || {} );
-	const dropDownMenuIcon = areAllOptionalControlsHidden ? plus : moreVertical;
 	const dropDownMenuLabelText = sprintf(
 		// translators: %s: The name of the tool e.g. "Color" or "Typography".
 		_x( '%s options', 'Button label to reveal tool panel options' ),
-		labelText
-	);
-	const dropdownMenuDescriptionText = areAllOptionalControlsHidden
-		? __( 'All options are currently hidden' )
-		: undefined;
-
-	const canResetAll = [ ...defaultItems, ...optionalItems ].some(
-		( [ , isSelected ] ) => isSelected
+		label
 	);
 
 	return (
 		<HStack { ...headerProps } ref={ forwardedRef }>
 			<Heading level={ 2 } className={ headingClassName }>
-				{ labelText }
+				{ label }
 			</Heading>
-			{ hasMenuItems && (
-				<DropdownMenu
-					icon={ dropDownMenuIcon }
-					label={ dropDownMenuLabelText }
-					menuProps={ { className: dropdownMenuClassName } }
-					toggleProps={ {
-						isSmall: true,
-						describedBy: dropdownMenuDescriptionText,
-					} }
-				>
-					{ () => (
-						<>
-							<DefaultControlsGroup
-								items={ defaultItems }
-								toggleItem={ toggleItem }
-							/>
-							<OptionalControlsGroup
-								items={ optionalItems }
-								toggleItem={ toggleItem }
-							/>
-							<MenuGroup>
-								<MenuItem
-									aria-disabled={ ! canResetAll }
-									variant={ 'tertiary' }
-									onClick={ () => {
-										if ( canResetAll ) {
-											resetAll();
-											speak(
-												__( 'All options reset' ),
-												'assertive'
-											);
-										}
-									} }
-								>
-									{ __( 'Reset all' ) }
-								</MenuItem>
-							</MenuGroup>
-						</>
-					) }
-				</DropdownMenu>
-			) }
+			<ToolsPanelDropdownMenu
+				label={ dropDownMenuLabelText }
+				resetAll={ resetAll }
+				toggleItem={ toggleItem }
+			/>
 		</HStack>
 	);
 };
