+++ packages/components/src/tools-panel/tools-panel-header/hook.ts
@@ -7,7 +7,6 @@ import { useMemo } from '@wordpress/element';
  * Internal dependencies
  */
 import * as styles from '../styles';
-import { useToolsPanelContext } from '../context';
 import { useContextSystem, WordPressComponentProps } from '../../ui/context';
 import { useCx } from '../../utils/hooks/use-cx';
 import type { ToolsPanelHeaderProps } from '../types';
@@ -25,24 +24,13 @@ export function useToolsPanelHeader(
 		return cx( styles.ToolsPanelHeader, className );
 	}, [ className, cx ] );
 
-	const dropdownMenuClassName = useMemo( () => {
-		return cx( styles.DropdownMenu );
-	}, [ cx ] );
-
 	const headingClassName = useMemo( () => {
 		return cx( styles.ToolsPanelHeading );
 	}, [ cx ] );
 
-	const { menuItems, hasMenuItems, areAllOptionalControlsHidden } =
-		useToolsPanelContext();
-
 	return {
 		...otherProps,
-		areAllOptionalControlsHidden,
-		dropdownMenuClassName,
-		hasMenuItems,
 		headingClassName,
-		menuItems,
 		className: classes,
 	};
 }
